Don't pass real minImageCount if it is less than 2
Some hardware (e.g., AMD Integrated Radeon Graphics) reports
capabilities.minImageCount of 1 from the physical device. ImGui throws
an assert in Debug mode if the value is less than 2, and for no good
reason: it has nothing to do with the functioning of ImGui. It might
be used by the ImGui utility functions for opening a window, which we
do not care about.

Eventually this should be a bug report against ImGui itself, but I
don't know the code well enough to make a pull request.